% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test Data for NPSS Scoring}
\format{
\subsection{\code{test_data}}{

A data frame with 10 rows and 29 columns:
\describe{
\item{npss_n}{NPSS Item Randomized Value}
...
}
}
}
\source{
https://pubmed.ncbi.nlm.nih.gov/36107710/
}
\usage{
test_data
}
\description{
Randomly sampled values within the range of possible responses to each item. Used for testing.
}
\keyword{datasets}
