% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_npss.R
\name{score_npss}
\alias{score_npss}
\title{score_npss}
\usage{
score_npss(input_df, missing_threshold = 0.5)
}
\arguments{
\item{input_df}{Dataframe containing the 29 NPSS items, with the prefix: 'npss'}

\item{missing_threshold}{A percentage of allowed missing data for each subscale in the NPSS (Default .5). If missingness exceeds the set threshold, function returns NA for subscale score.}
}
\value{
Input dataframe with NPSS scores added.
}
\description{
score_npss
}
